// ecos/gui/apps/renderer/src/api/type.ts
export enum CMDEnum {
  create_workspace = "create_workspace",
  load_workspace = "load_workspace",
  rtl2gds = "rtl2gds",
  run_step = "run_step",
  get_info = "get_info",
  home_page = "home_page"
}

// get_info command 的 id 枚举
export enum InfoEnum {
  views = "views",
  layout = "layout",
  metrics = "metrics",
  subflow = "subflow",
  analysis = "analysis",
  maps = "maps",
  checklist = "checklist",
  sta = "sta",
  config = "config"
}

export enum StepEnum {
  RTL2GDS = "RTL2GDS",
  INIT = "Init",
  SOC = "SOC",
  FUNC_SIM = "Functional simulation",
  SYNTHESIS = "Synthesis",
  FLOORPLAN = "Floorplan",
  NETLIST_OPT = "fixFanout",
  PLACEMENT = "place",
  CTS = "CTS",
  PNP = "PNP",
  TIMING_OPT = "Timing optimization",
  TIMING_OPT_DRV = "optDrv",
  TIMING_OPT_HOLD = "optHold",
  TIMING_OPT_SETUP = "optSetup",
  LEGALIZATION = "legalization",
  ROUTING = "route",
  FILLER = "filler",
  GDS = "GDS",
  SIGNOFF = "Signoff",
  STA = "sta",
  DRC = "drc",
  RCX = "RCX",
  ABSTRACT_LEF = "Abstract lef",
  MERGE = "GDS merge"
}
// ...
export enum ResponseEnum {
  success = "success",
  failed = "failed",
  error = "error",
  warning = "warning"
}

export enum StateEnum {
  Invalid = "Invalid",
  Unstart = "Unstart",
  Success = "Success",
  Ongoing = "Ongoing",
  Pending = "Pending",
  Imcomplete = "Imcomplete",
  // Ignored = "Ignored",
}

export enum CheckState {
  Unstart = "Unstart",
  Success = "Success",
  Failed = "Failed",
  Warning = "Warning",
}

export interface RequestData<T> {
  cmd: CMDEnum;
  data: T;
}

export interface ResponseData<T> {
  cmd: CMDEnum;
  response: ResponseEnum;
  data: T;
  message: string[];
}